requirements.txt
numpy>=1.18.2
pytesseract>=0.3.4
opencv-python>=4.3.0.34
Pillow>=7.1.1
PyAutoGUI>=0.9.50